backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
motor==3.6.0
pymongo==4.9.1
pydantic==2.7.1
pydantic-settings==2.3.3
python-multipart==0.0.9
python-dotenv==1.0.1
httpx==0.27.0
PyYAML==6.0.1
aiofiles==23.2.1
openai==1.35.3
tenacity==8.3.0
pytest==8.2.2
pytest-asyncio==0.23.7
